+++ docs/CHANGELOG.md
@@ -3,6 +3,12 @@ Doom Struct (C) 2015-2020
 by Matt Tropiano et al. (see AUTHORS.txt)
 
 
+Changed in [NOW]
+----------------
+
+- `Added` TextureUtils.TextureCopier methods: copyTexture() with replace, ignoreFlat().
+
+
 Changed in 2.9.1
 ----------------